+++ lib/praxis.rb
@@ -50,7 +50,9 @@ module Praxis
 
   # types
   module Types
+    autoload :FuzzyHash, 'praxis/types/fuzzy_hash'
     autoload :MediaTypeCommon, 'praxis/types/media_type_common'
+    autoload :MultipartArray, 'praxis/types/multipart_array'
   end
 
   autoload :Links, 'praxis/links'
@@ -80,6 +82,7 @@ module Praxis
 
 
   module Handlers
+    autoload :Plain, 'praxis/handlers/plain'
     autoload :JSON, 'praxis/handlers/json'
     autoload :WWWForm, 'praxis/handlers/www_form'
     autoload :XML, 'praxis/handlers/xml'
@@ -113,5 +116,6 @@ module Praxis
   require 'praxis/responses/http'
   require 'praxis/responses/internal_server_error'
   require 'praxis/responses/validation_error'
+  require 'praxis/responses/multipart_ok'
 
 end