+++ spec/spec_app/app/responses/bulk_response.rb
@@ -1,6 +1,0 @@
-
-class BulkResponse < Praxis::Response
-  self.response_name = :bulk_response
-  self.status = 200
-end
-